Organized svg icons on the components folder
